Extract union attribute generation
Move union attribue generation to its own namespace.
Reorganize other namespaces.
